Refactor App folder
