+++ src/dashboard-react-web/src/components/App/AccountDropdown/AccountDropdown.scss
@@ -1,4 +1,4 @@
-@use '../../styles/fonts.scss';
+@use '../../../styles/fonts.scss';
 
 .account-dropdown {
   font-family: fonts.$monospace-font-family;
