+++ src/dashboard-react-web/src/components/App/AccountDropdown/index.ts
@@ -1,0 +1,1 @@
+export { AccountDropDown, AccountDropDownPure } from './AccountDropdown';
