+++ src/dashboard-react-web/src/components/App/Header/Header.scss
@@ -1,4 +1,4 @@
-@use '../../styles/layout-variables.scss';
+@use '../../../styles/layout-variables.scss';
 
 $-title-font-weight: 500;
