+++ src/dashboard-react-web/src/components/App/Header/index.ts
@@ -1,0 +1,1 @@
+export { Header, HeaderPure } from './Header';
