+++ src/dashboard-react-web/src/components/App/NavBar/NavBar.scss
@@ -1,5 +1,5 @@
-@use "../../styles/layout-variables.scss";
-@use "../../styles/ant-colors.scss";
+@use "../../../styles/layout-variables.scss";
+@use "../../../styles/ant-colors.scss";
 
 .navbar {
   position: fixed;
