+++ src/dashboard-react-web/src/components/App/NavBar/index.ts
@@ -1,0 +1,1 @@
+export { NavBar, NavBarPure } from './NavBar';
